chore: Update background color in app.component.scss and user-management.component.scss
This commit updates the background color in the `app.component.scss` file from `#f5f5f5` to `#000000` to improve the visual appearance. It also adds a new background color to the `.wrapper` class in the `user-management.component.scss` file, setting it to `gray` for better contrast.

Note: The recent user commits and repository commits do not provide any additional relevant information for generating the commit message.