chore: Update cyber-menu component imports and styles
This commit updates the imports in the cyber-menu.component.ts file to include the CommonModule and RouterLink from the @angular/common and @angular/router respectively. It also updates the styles in the cyber-menu.component.scss file by importing the colors, mixins, and variables from the styles folder.

Note: The recent user commits and repository commits do not provide any additional relevant information for generating the commit message.