chore: Update app.config.ts to provide HttpClient with Fetch
This commit updates the app.config.ts file to import the provideHttpClient and withFetch functions from the @angular/common/http module. It then adds the provideHttpClient(withFetch()) function to the list of providers in the appConfig constant. This change enables the use of the Fetch API for HTTP requests in the application.

Note: The recent user commits and repository commits do not provide any additional relevant information for generating the commit message.