chore: Update app.component.html, app.component.ts, toolbar.component.html, and db.json
This commit updates the app.component.html file to comment out the <app-user-management> component and add a <router-outlet> and <app-component-documentation> component. It also updates the app.component.ts file to import the ComponentDocumentationComponent. Additionally, it comments out the <app-cyber-menu> component in the toolbar.component.html file. Finally, it adds a new user object to the db.json file.