chore: Update comment component to display comments and allow adding new comments
This commit updates the comment component to display a list of comments and allows users to add new comments. The HTML template has been modified to iterate over the comments array and display each comment with the user's profile picture, name, and text. The SCSS file has been updated to style the comment container, comment body, and add comment section. The TypeScript file now includes an input property to receive the comments array and a newCommentText property to store the text of the new comment. The addComment() method has been implemented to add a new comment to the comments array when the "Post" button is clicked. This change enhances the functionality of the comment component by enabling users to view and interact with comments.

Note: The recent user commits and repository commits do not provide any additional relevant information for generating the commit message.