chore: Update PostService to use HttpClient for API requests
This commit updates the PostService to use the HttpClient module from @angular/common/http for making API requests. The getPosts() method now sends a GET request to the specified API endpoint to fetch the list of posts. Additionally, a new createPost() method has been added to send a POST request for creating a new post. This change improves the functionality of the PostService by utilizing the HttpClient module for efficient and standardized API communication.