chore: Include app-notification and app-create-post components in quickfeed.component.html
This commit updates the quickfeed.component.html file to include the app-notification and app-create-post components. The commented out lines for these components have been uncommented, allowing them to be rendered in the feed-container. This change enhances the functionality of the QuickfeedComponent by adding the ability to display notifications and create new posts.